Refactor App component and implement separate forms for configuration, vendor, and customer counts
